`@remotion/browser-studio`: Support deleting canvas items
